changed input casting

+++ Lab2/RoomCalculator/Program.cs
@@ -13,11 +13,11 @@ namespace RoomCalculator
             do {
                 // prompt the user for input
                 Console.Write("Enter Length:");
-                double l = int.Parse(Console.ReadLine());
+                double l = double.Parse(Console.ReadLine());
                 Console.Write("Enter Width:");
-                double w = int.Parse(Console.ReadLine());
+                double w = double.Parse(Console.ReadLine());
                 Console.Write("Enter Height:");
-                double h = int.Parse(Console.ReadLine());
+                double h = double.Parse(Console.ReadLine());
 
                 // calculate values and output results
                 Console.WriteLine("Area:" + l * w);
